backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.49
psycopg2-binary==2.9.6
python-dotenv==1.0.0
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
google-generativeai==0.3.0
packaging==23.0
alembic==1.12.1
